requirements.txt
Django==1.10.4
django-filter==1.0.1
django-mathfilters==0.4.0
django-registration-redux==1.4
djangorestframework==3.5.3
lxml==3.7.3
Pillow==4.0.0
python-dateutil==1.5
pytz==2013.7
requests==2.12.5
